代码重构
- 添加dp和px之间相互转换的方法
- 把工具类移到sdk目录下
