+++ appLibrary/src/main/java/com/android/lvxin/sdk/common/DisplayUtils.java
@@ -1,0 +1,43 @@
+package com.android.lvxin.sdk.common;
+
+import android.content.Context;
+import android.util.TypedValue;
+
+/**
+ * @ClassName: DisplayUtils
+ * @Description: 常用单位转换的辅助类
+ * @Author: lvxin
+ * @Date: 12/1/15 09:38
+ */
+public class DisplayUtils {
+    /**
+     * dp转px
+     */
+    public static int dp2px(Context context, float dpVal) {
+        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
+                dpVal, context.getResources().getDisplayMetrics());
+    }
+
+    /**
+     * sp转px
+     */
+    public static int sp2px(Context context, float spVal) {
+        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
+                spVal, context.getResources().getDisplayMetrics());
+    }
+
+    /**
+     * px转dp
+     */
+    public static float px2dp(Context context, float pxVal) {
+        final float scale = context.getResources().getDisplayMetrics().density;
+        return (pxVal / scale);
+    }
+
+    /**
+     * px转sp
+     */
+    public static float px2sp(Context context, float pxVal) {
+        return (pxVal / context.getResources().getDisplayMetrics().scaledDensity);
+    }
+}
